% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbmf.R
\name{evb_global_pca}
\alias{evb_global_pca}
\title{Iterative EVB analytic solution by Nakajima et al. 2015}
\usage{
evb_global_pca(svd_result, L, M, H, sigma2 = NULL)
}
\description{
Iterative EVB analytic solution by Nakajima et al. 2015
}
