% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbmf.R
\name{evb_ite_pca}
\alias{evb_ite_pca}
\title{EVB analytic global solution by Nakajima et al. 2015}
\usage{
evb_ite_pca(svd_result, L, M, max_iter = 100, tol = 1e-06, local = TRUE)
}
\description{
EVB analytic global solution by Nakajima et al. 2015
}
