% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpca.R
\name{get.alpha.mle}
\alias{get.alpha.mle}
\title{Get MLE estimate of alpha}
\usage{
get.alpha.mle(m.W, Sigma.W)
}
\arguments{
\item{m.W}{\eqn{\mathbb{E}[\mathbf{W}]}.}

\item{Sigma.W}{\eqn{\Sigma_{\mathbf{W}}}.}
}
\value{
\eqn{\hat{\alpha}_{\text{MLE}}} of \eqn{P(W \mid \alpha)}
}
\description{
Get MLE estimate of alpha
}
