% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpca.R
\name{get.alpha}
\alias{get.alpha}
\title{Title Update \eqn{Q(\mathbf{\alpha})}.}
\usage{
get.alpha(m.W, Sigma.W, a.alpha, b.alpha)
}
\arguments{
\item{m.W}{\eqn{\mathbb{E}[\mathbf{W}]}.}

\item{Sigma.W}{\eqn{\Sigma_{\mathbf{W}}}.}

\item{a.alpha}{\eqn{a_\alpha}.}

\item{b.alpha}{\eqn{b_\alpha}.}
}
\value{
\eqn{\tilde{a}_\alpha} and \eqn{\tilde{b}_\alpha}.
}
\description{
Title Update \eqn{Q(\mathbf{\alpha})}.
}
