% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpca.R
\name{get.mu}
\alias{get.mu}
\title{Update \eqn{Q(\mathbf{\mu})}.}
\usage{
get.mu(t, m.W, m.x, m.tau, beta)
}
\arguments{
\item{t}{d-by-n data matrix.}

\item{m.W}{\eqn{\mathbb{E}_Q[\mathbf{W}]}.}

\item{m.x}{\eqn{\mathbb{E}[\mathbf{x}_n]}.}

\item{m.tau}{\eqn{\mathbb{E}_Q[\tau]}.}

\item{beta}{\eqn{\beta}.}
}
\value{
\eqn{\mathbb{E}[\mu]} and \eqn{\Sigma_\mu}.
}
\description{
Update \eqn{Q(\mathbf{\mu})}.
}
