% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpca.R
\name{get.tau}
\alias{get.tau}
\title{Title Update \eqn{Q(\tau)}.}
\usage{
get.tau(t, m.x, Sigma.x, m.W, Sigma.W, m.mu, Sigma.mu, a.tau, b.tau)
}
\arguments{
\item{t}{d-by-n data matrix.}

\item{m.x}{\eqn{\mathbb{E}[\mathbf{x}_n]}.}

\item{Sigma.x}{\eqn{\Sigma_\mathbf{x}}.}

\item{m.W}{\eqn{\mathbb{E}[\mathbf{W}]}.}

\item{Sigma.W}{\eqn{\Sigma_{\mathbf{W}}}.}

\item{m.mu}{\eqn{\mathbb{E}_Q[\mu]}.}

\item{Sigma.mu}{\eqn{\Sigma_{\mathbf{\mu}}}.}

\item{a.tau}{\eqn{a_\tau}.}

\item{b.tau}{\eqn{b_\tau}.}
}
\value{
\eqn{\tilde{a}_\tau} and \eqn{\tilde{b}_\tau}.
}
\description{
Title Update \eqn{Q(\tau)}.
}
