% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpca.R
\name{get.W}
\alias{get.W}
\title{Title Update \eqn{\mathbf{W}}.}
\usage{
get.W(t, m.x, Sigma.x, m.mu, m.tau, m.alpha)
}
\arguments{
\item{t}{d-by-n data matrix.}

\item{m.x}{\eqn{\mathbb{E}[\mathbf{x}_n]}.}

\item{Sigma.x}{\eqn{\Sigma_\mathbf{x}}.}

\item{m.mu}{\eqn{\mathbb{E}[\mu]}.}

\item{m.tau}{\eqn{\mathbb{E}_Q[\tau]}.}

\item{m.alpha}{\eqn{\mathbb{E}[\mathbf{\alpha}]}.}
}
\value{
\eqn{\mathbb{E}[\mathbf{W}]} and \eqn{\Sigma_{\mathbf{W}}}.
}
\description{
Title Update \eqn{\mathbf{W}}.
}
