% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpca.R
\name{get.x}
\alias{get.x}
\title{Title Update \eqn{Q(X)}.}
\usage{
get.x(t, m.W, Sigma.W, m.mu, m.tau)
}
\arguments{
\item{t}{d-by-n data matrix.}

\item{m.W}{\eqn{\mathbb{E}_Q[\mathbf{W}]}.}

\item{Sigma.W}{\eqn{\Sigma_\mathbf{W}}.}

\item{m.mu}{\eqn{\mathbb{E}_Q[\mu]}.}

\item{m.tau}{\eqn{\mathbb{E}_Q[\tau]}.}
}
\value{
\eqn{\mathbb{E}[\mathbf{x}_n]} and \eqn{\Sigma_\mathbf{x}}.
}
\description{
Title Update \eqn{Q(X)}.
}
