% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpca_draft.R
\name{update_alpha}
\alias{update_alpha}
\title{Update \eqn{Q(\mathbf{\alpha_\mathbf{X}})}.}
\usage{
update_alpha(obj)
}
\arguments{
\item{obj}{vpca.data object.}
}
\value{
vpca.data object.
}
\description{
Update \eqn{Q(\mathbf{\alpha_\mathbf{X}})}.
}
