% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_flashier_F.R
\name{update_flashier_F}
\alias{update_flashier_F}
\title{Update factor F in a flash object assuming iid N(0,1) prior}
\usage{
update_flashier_F(flash_obj)
}
\arguments{
\item{flash_obj}{flash object}
}
\value{
flash object with updated EF EF2
}
\description{
Update factor F in a flash object assuming iid N(0,1) prior
}
