% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpca_draft.R
\name{update_mu}
\alias{update_mu}
\title{Update \eqn{Q(\mathbf{\mu})}.}
\usage{
update_mu(obj)
}
\arguments{
\item{obj}{vpca.data object.}
}
\value{
vpca.data object.
}
\description{
Update \eqn{Q(\mathbf{\mu})}.
}
