% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpca_draft.R
\name{update_X}
\alias{update_X}
\title{Update either \eqn{Q(\mathbf{X})}.}
\usage{
update_X(obj)
}
\arguments{
\item{obj}{vpca.data object.}
}
\value{
vpca.data object.
}
\description{
Update either \eqn{Q(\mathbf{X})}.
}
