% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpca.R
\name{vpcar}
\alias{vpcar}
\title{Variational PCA algorithm runner.}
\usage{
vpcar(init, min.iter = 1000, max.iter = 3000, epsilon = 1e-04)
}
\arguments{
\item{init}{vpca.data object.}

\item{min.iter}{Minimum number of iterations.}

\item{max.iter}{Maximum number of iterations.}

\item{epsilon}{The algorithm terminates if the difference in ELBO is less than \eqn{\epsilon}.}
}
\value{
First moments and ELBOs in all iterations.
}
\description{
Variational PCA algorithm runner.
}
