% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpca.R
\name{get.elbo}
\alias{get.elbo}
\title{Compute ELBO with given moments,}
\usage{
get.elbo(
  t,
  m.x,
  Sigma.x,
  m.mu,
  Sigma.mu,
  m.W,
  Sigma.W,
  alpha.param,
  a.tilde.tau,
  b.tilde.tau,
  a.tau,
  b.tau,
  beta
)
}
\value{
ELBO \eqn{\mathcal{L}(\theta)}.
}
\description{
Compute ELBO with given moments,
}
