% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpca.R
\name{vpca}
\alias{vpca}
\title{Variational PCA algorithm runner.}
\usage{
vpca(
  t,
  fit.alpha.mle = F,
  W.alpha = T,
  x.alpha = F,
  a.alpha = 0.001,
  b.alpha = 0.001,
  a.tau = 0.001,
  b.tau = 0.001,
  beta = 0.001,
  min.iter = 1000,
  max.iter = 3000,
  epsilon = 1e-04
)
}
\arguments{
\item{t}{d-by-n data matrix.}

\item{fit.alpha.mle}{Whether to fit \eqn{\alpha} via MLE.}

\item{W.alpha}{Whether to fit \eqn{\alpha} for \eqn{W}.}

\item{x.alpha}{Whether to fit \eqn{\alpha} for \eqn{x}.}

\item{a.alpha}{\eqn{a_\alpha}.}

\item{b.alpha}{\eqn{b_\alpha}.}

\item{a.tau}{\eqn{a_\tau}.}

\item{b.tau}{\eqn{a_\alpha}.}

\item{beta}{\eqn{\beta}.}

\item{min.iter}{Minimum number of iterations.}

\item{max.iter}{Maximum number of iterations.}

\item{epsilon}{The algorithm terminates if the difference in ELBO is less than \eqn{\epsilon}.}
}
\value{
\eqn{\mathbb{E}[W]} and ELBOs in all iterations.
}
\description{
Variational PCA algorithm runner.
}
